+++ test/e2e/testdata/curl.yaml
@@ -2,7 +2,6 @@ apiVersion: apps/v1
 kind: Deployment
 metadata:
   name: curl
-  namespace: {{.Namespace}}
 spec:
   replicas: 1
   selector:
