+++ test/e2e/testdata/helloworld-service.yaml
@@ -2,7 +2,6 @@ apiVersion: v1
 kind: Service
 metadata:
   name: helloworld
-  namespace: {{.Namespace}}
   labels:
     app: helloworld
     service: helloworld
